pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "MEGAFON local community platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "sqlmodel>=0.0.24",
]
